Fix for bug #160766829 - wait until frag started
We must wait until NavHostFragment has cycled back to its started state before calling navigate to navigate to a URI in order for arguments to be passed to root destinations.
Here, we're doing so by making use of the Lifecycle extensions to launch a coroutine which will fire when the NavHostFragment has been started again.